fix: Permission cleanup and UX fixes
