+++ frappe/core/doctype/allowed_module/allowed_module.json
@@ -1,0 +1,31 @@
+{
+ "actions": [],
+ "creation": "2020-12-22 21:53:53.309663",
+ "doctype": "DocType",
+ "editable_grid": 1,
+ "engine": "InnoDB",
+ "field_order": [
+  "module"
+ ],
+ "fields": [
+  {
+   "fieldname": "module",
+   "fieldtype": "Link",
+   "label": "Module",
+   "options": "Module Def"
+  }
+ ],
+ "index_web_pages_for_search": 1,
+ "istable": 1,
+ "links": [],
+ "modified": "2020-12-22 22:10:20.346492",
+ "modified_by": "Administrator",
+ "module": "Core",
+ "name": "Allowed Module",
+ "owner": "Administrator",
+ "permissions": [],
+ "quick_entry": 1,
+ "sort_field": "modified",
+ "sort_order": "DESC",
+ "track_changes": 1
+}
