+++ frappe/core/doctype/module_profile/module_profile.js
@@ -1,0 +1,15 @@
+// Copyright (c) 2020, Frappe Technologies and contributors
+// For license information, please see license.txt
+
+frappe.ui.form.on('Module Profile', {
+	onload: function(frm) {
+		if(has_common(frappe.user_roles, ["Administrator", "System Manager"])) {
+			if(!frm.module_editor) {
+				let module_area = $('<div style="min-height: 300px">')
+					.appendTo(frm.fields_dict.module_html.wrapper);
+
+				frm.module_editor = new frappe.ModuleEditor(frm, module_area);
+			}
+		}
+	}
+});
