+++ frappe/core/doctype/module_profile/module_profile.json
@@ -1,0 +1,57 @@
+{
+ "actions": [],
+ "creation": "2020-12-22 22:00:30.614475",
+ "doctype": "DocType",
+ "editable_grid": 1,
+ "engine": "InnoDB",
+ "field_order": [
+  "module_profile_name",
+  "module_html",
+  "modules"
+ ],
+ "fields": [
+  {
+   "fieldname": "module_profile_name",
+   "fieldtype": "Data",
+   "in_list_view": 1,
+   "label": "Module Profile Name",
+   "reqd": 1
+  },
+  {
+   "fieldname": "module_html",
+   "fieldtype": "HTML",
+   "label": "Module HTML"
+  },
+  {
+   "fieldname": "modules",
+   "fieldtype": "Table",
+   "label": "Module Assigned",
+   "options": "Allowed Module"
+  }
+ ],
+ "index_web_pages_for_search": 1,
+ "links": [],
+ "modified": "2020-12-22 22:10:00.717009",
+ "modified_by": "Administrator",
+ "module": "Core",
+ "name": "Module Profile",
+ "owner": "Administrator",
+ "permissions": [
+  {
+   "create": 1,
+   "delete": 1,
+   "email": 1,
+   "export": 1,
+   "print": 1,
+   "read": 1,
+   "report": 1,
+   "role": "System Manager",
+   "share": 1,
+   "write": 1
+  }
+ ],
+ "quick_entry": 1,
+ "sort_field": "modified",
+ "sort_order": "DESC",
+ "track_changes": 1
+}
