+++ frappe/core/doctype/module_profile/test_module_profile.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2020, Frappe Technologies and Contributors
+# See license.txt
+from __future__ import unicode_literals
+
+# import frappe
+import unittest
+
+class TestModuleProfile(unittest.TestCase):
+	pass
